Have HandlerInvoker keep a reference to target instance

+++ common/src/main/java/io/fluxcapacitor/common/api/tracking/ReadAction.java
@@ -26,9 +26,9 @@ public class ReadAction implements ClientAction {
     String processor;
     int channel;
     int maxSize;
-    int maxTimeout;
+    long maxTimeout;
 
     int size;
     Long lastIndex;
-    int timeout;
+    long timeout;
 }
